refactor: split oversized files (sandbox, tools/mod)
Split sandbox.rs (1,452 lines) into sandbox/mod.rs + bwrap.rs +
macos.rs + docker.rs. Split tools/mod.rs into mod.rs + registry.rs.

Public API unchanged -- all existing imports continue to work.
Tests moved to their respective backend modules.

Closes #194
