+++ crates/octos-agent/src/sandbox/bwrap.rs
@@ -1,0 +1,116 @@
+//! Linux sandbox using bubblewrap (bwrap).
+
+use std::path::Path;
+
+use tokio::process::Command;
+
+use super::{Sandbox, BLOCKED_ENV_VARS};
+
+/// Linux sandbox using bubblewrap (bwrap).
+pub struct BwrapSandbox {
+    pub(crate) allow_network: bool,
+}
+
+impl Sandbox for BwrapSandbox {
+    fn wrap_command(&self, shell_command: &str, cwd: &Path) -> Command {
+        let mut cmd = Command::new("bwrap");
+
+        // Clear dangerous environment variables before entering sandbox
+        for var in BLOCKED_ENV_VARS {
+            cmd.env_remove(var);
+        }
+
+        // Read-only bind system directories
+        for dir in &["/usr", "/lib", "/lib64", "/bin", "/sbin", "/etc"] {
+            if Path::new(dir).exists() {
+                cmd.arg("--ro-bind").arg(dir).arg(dir);
+            }
+        }
+
+        // Read-write bind the working directory
+        let cwd_str = cwd.to_string_lossy();
+        cmd.arg("--bind").arg(&*cwd_str).arg(&*cwd_str);
+
+        // Bind /tmp for scratch space
+        cmd.arg("--tmpfs").arg("/tmp");
+
+        // /dev minimal
+        cmd.arg("--dev").arg("/dev");
+        cmd.arg("--proc").arg("/proc");
+
+        if !self.allow_network {
+            cmd.arg("--unshare-net");
+        }
+
+        cmd.arg("--unshare-pid");
+        cmd.arg("--die-with-parent");
+        cmd.arg("--chdir").arg(&*cwd_str);
+        cmd.arg("--").arg("sh").arg("-c").arg(shell_command);
+
+        cmd
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_bwrap_sandbox_command() {
+        let sb = BwrapSandbox {
+            allow_network: false,
+        };
+        let cmd = sb.wrap_command("echo hi", Path::new("/tmp/test"));
+        let prog = cmd.as_std().get_program().to_string_lossy().to_string();
+        assert_eq!(prog, "bwrap");
+        let args: Vec<_> = cmd
+            .as_std()
+            .get_args()
+            .map(|a| a.to_string_lossy().to_string())
+            .collect();
+        assert!(args.contains(&"--unshare-net".to_string()));
+        assert!(args.contains(&"echo hi".to_string()));
+    }
+
+    #[test]
+    fn test_bwrap_sandbox_env_sanitization() {
+        let sb = BwrapSandbox {
+            allow_network: false,
+        };
+        let cmd = sb.wrap_command("ls", Path::new("/tmp"));
+        let removed: Vec<String> = cmd
+            .as_std()
+            .get_envs()
+            .filter_map(|(k, v)| {
+                if v.is_none() {
+                    Some(k.to_string_lossy().to_string())
+                } else {
+                    None
+                }
+            })
+            .collect();
+        for var in BLOCKED_ENV_VARS {
+            assert!(
+                removed.iter().any(|r| r == *var),
+                "bwrap should env_remove {var}"
+            );
+        }
+    }
+
+    #[test]
+    fn test_bwrap_sandbox_allows_network() {
+        let sb = BwrapSandbox {
+            allow_network: true,
+        };
+        let cmd = sb.wrap_command("echo hi", Path::new("/tmp"));
+        let args: Vec<_> = cmd
+            .as_std()
+            .get_args()
+            .map(|a| a.to_string_lossy().to_string())
+            .collect();
+        assert!(
+            !args.contains(&"--unshare-net".to_string()),
+            "should not unshare net when network is allowed"
+        );
+    }
+}
