+++ crates/octos-agent/src/tools/mod.rs
@@ -7,7 +7,6 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use eyre::Result;
 use octos_core::TokenUsage;
-use octos_llm::ToolSpec;
 
 use crate::progress::ProgressReporter;
 
@@ -88,15 +87,15 @@ pub trait Tool: Send + Sync {
 /// exceeds a threshold. Base tools are pinned and never evicted.
 pub struct ToolLifecycle {
     /// Per-tool last-used iteration counter.
-    last_used: HashMap<String, u32>,
+    pub(crate) last_used: HashMap<String, u32>,
     /// Current iteration counter.
-    iteration: u32,
+    pub(crate) iteration: u32,
     /// Tools that are never auto-evicted.
-    base_tools: HashSet<String>,
+    pub(crate) base_tools: HashSet<String>,
     /// Maximum active tools before eviction kicks in.
-    max_active: usize,
+    pub(crate) max_active: usize,
     /// Tools idle for this many iterations become eviction candidates.
-    idle_threshold: u32,
+    pub(crate) idle_threshold: u32,
 }
 
 impl Default for ToolLifecycle {
@@ -161,469 +160,9 @@ impl ToolLifecycle {
     }
 }
 
-/// Registry of available tools.
-pub struct ToolRegistry {
-    tools: HashMap<String, Arc<dyn Tool>>,
-    /// Provider-specific policy that filters specs() output without removing tools.
-    provider_policy: Option<ToolPolicy>,
-    /// Context-based tag filter: only tools with matching tags appear in specs().
-    /// Tools with empty tags always pass.
-    context_filter: Option<Vec<String>>,
-    /// Cached specs output, invalidated on registry mutations.
-    cached_specs: std::sync::Mutex<Option<Vec<ToolSpec>>>,
-    /// Deferred tools: registered but hidden from specs() until activated.
-    /// Uses interior mutability so activate() can work through Arc<ToolRegistry>.
-    deferred: std::sync::Mutex<HashSet<String>>,
-    /// LRU lifecycle manager for auto-eviction of idle tools.
-    lifecycle: std::sync::Mutex<ToolLifecycle>,
-    /// Tool names that came from plugin binaries (for auto-send hook filtering).
-    plugin_tools: HashSet<String>,
-    /// Tools that are permanently deferred (spawn_only) — cannot be activated by activate_tools.
-    spawn_only: HashSet<String>,
-    /// Custom messages for spawn_only tools returned to the LLM after auto-backgrounding.
-    spawn_only_messages: HashMap<String, String>,
-}
-
-impl Default for ToolRegistry {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl ToolRegistry {
-    /// Create an empty tool registry.
-    pub fn new() -> Self {
-        Self {
-            tools: HashMap::new(),
-            provider_policy: None,
-            context_filter: None,
-            cached_specs: std::sync::Mutex::new(None),
-            deferred: std::sync::Mutex::new(HashSet::new()),
-            lifecycle: std::sync::Mutex::new(ToolLifecycle::default()),
-            plugin_tools: HashSet::new(),
-            spawn_only: HashSet::new(),
-            spawn_only_messages: HashMap::new(),
-        }
-    }
-
-    /// Mark a tool name as coming from a plugin binary.
-    pub fn mark_as_plugin(&mut self, name: &str) {
-        self.plugin_tools.insert(name.to_string());
-    }
-
-    /// Mark a tool as spawn_only with an optional custom message.
-    pub fn mark_spawn_only(&mut self, name: &str, message: Option<String>) {
-        self.spawn_only.insert(name.to_string());
-        if let Some(msg) = message {
-            self.spawn_only_messages.insert(name.to_string(), msg);
-        }
-    }
-
-    /// Check if a tool is marked spawn_only.
-    pub fn is_spawn_only(&self, name: &str) -> bool {
-        self.spawn_only.contains(name)
-    }
-
-    /// Get the custom message for a spawn_only tool, or a default.
-    pub fn spawn_only_message(&self, name: &str) -> String {
-        self.spawn_only_messages
-            .get(name)
-            .cloned()
-            .unwrap_or_else(|| "SUCCESS: Task is now running in background. The result will be delivered to the user automatically. No further action needed.".to_string())
-    }
-
-    /// Check if a tool came from a plugin binary.
-    pub fn is_plugin(&self, name: &str) -> bool {
-        self.plugin_tools.contains(name)
-    }
-
-    /// Register a tool.
-    pub fn register(&mut self, tool: impl Tool + 'static) {
-        self.tools.insert(tool.name().to_string(), Arc::new(tool));
-        self.invalidate_cache();
-    }
-
-    /// Register a tool from an existing Arc (for keeping a separate reference).
-    pub fn register_arc(&mut self, tool: Arc<dyn Tool>) {
-        self.tools.insert(tool.name().to_string(), tool);
-        self.invalidate_cache();
-    }
-
-    /// Get tool specifications for the LLM, filtered by provider policy if set.
-    /// Results are cached and invalidated when the registry is mutated.
-    pub fn specs(&self) -> Vec<ToolSpec> {
-        let mut cache = self.cached_specs.lock().unwrap_or_else(|e| e.into_inner());
-        if let Some(ref specs) = *cache {
-            return specs.clone();
-        }
-
-        let deferred = self.deferred.lock().unwrap_or_else(|e| e.into_inner());
-        let specs: Vec<ToolSpec> = self
-            .tools
-            .values()
-            .filter(|t| !deferred.contains(t.name()))
-            .filter(|t| {
-                self.provider_policy
-                    .as_ref()
-                    .is_none_or(|p| p.is_allowed_with_tags(t.name(), t.tags()))
-            })
-            .filter(|t| {
-                self.context_filter.as_ref().is_none_or(|tags| {
-                    // Tools with no tags pass through; tools with tags must match
-                    let tool_tags = t.tags();
-                    tool_tags.is_empty()
-                        || tool_tags.iter().any(|tag| tags.contains(&tag.to_string()))
-                })
-            })
-            .map(|t| ToolSpec {
-                name: t.name().to_string(),
-                description: t.description().to_string(),
-                input_schema: t.input_schema(),
-            })
-            .collect();
-
-        *cache = Some(specs.clone());
-        specs
-    }
-
-    /// Get a tool by name.
-    pub fn get(&self, name: &str) -> Option<&Arc<dyn Tool>> {
-        self.tools.get(name)
-    }
-
-    /// Number of registered tools.
-    pub fn len(&self) -> usize {
-        self.tools.len()
-    }
-
-    /// Whether the registry is empty.
-    pub fn is_empty(&self) -> bool {
-        self.tools.is_empty()
-    }
-
-    /// Retain only tools whose names satisfy the predicate.
-    pub fn retain(&mut self, f: impl Fn(&str) -> bool) {
-        self.tools.retain(|name, _| f(name));
-        self.invalidate_cache();
-    }
-
-    /// Remove tools not permitted by the given policy.
-    pub fn apply_policy(&mut self, policy: &ToolPolicy) {
-        if policy.is_empty() {
-            return;
-        }
-        self.retain(|name| policy.is_allowed(name));
-    }
-
-    /// Set a provider-specific policy that filters `specs()` and `execute()`.
-    ///
-    /// Unlike `apply_policy` which permanently removes tools from the registry,
-    /// this keeps tools registered but blocks both spec visibility and execution.
-    pub fn set_provider_policy(&mut self, policy: ToolPolicy) {
-        if policy.is_empty() {
-            return;
-        }
-        self.provider_policy = Some(policy);
-        self.invalidate_cache();
-    }
-
-    /// Return the current provider policy (if any), so callers like SpawnTool
-    /// can propagate it to subagent registries.
-    pub fn provider_policy(&self) -> Option<&ToolPolicy> {
-        self.provider_policy.as_ref()
-    }
-
-    /// Set a context-based tag filter. Only tools whose tags overlap with these
-    /// values will appear in `specs()`. Tools with no tags always pass through.
-    pub fn set_context_filter(&mut self, tags: Vec<String>) {
-        if tags.is_empty() {
-            return;
-        }
-        self.context_filter = Some(tags);
-        self.invalidate_cache();
-    }
-
-    /// Create a new ToolRegistry by cloning all tools except the named exclusions.
-    ///
-    /// The new registry shares the same `Arc<dyn Tool>` instances (cheap).
-    /// Provider policy and context filter are also copied.
-    pub fn snapshot_excluding(&self, exclude: &[&str]) -> Self {
-        let tools: HashMap<String, Arc<dyn Tool>> = self
-            .tools
-            .iter()
-            .filter(|(name, _)| !exclude.contains(&name.as_str()))
-            .map(|(k, v)| (k.clone(), v.clone()))
-            .collect();
-        let deferred = self
-            .deferred
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .clone();
-        let parent = self.lifecycle.lock().unwrap_or_else(|e| e.into_inner());
-        let lifecycle = ToolLifecycle {
-            last_used: HashMap::new(),
-            iteration: 0,
-            base_tools: parent.base_tools.clone(),
-            max_active: parent.max_active,
-            idle_threshold: parent.idle_threshold,
-        };
-        drop(parent);
-
-        Self {
-            tools,
-            provider_policy: self.provider_policy.clone(),
-            context_filter: self.context_filter.clone(),
-            cached_specs: std::sync::Mutex::new(None),
-            deferred: std::sync::Mutex::new(deferred),
-            lifecycle: std::sync::Mutex::new(lifecycle),
-            plugin_tools: self.plugin_tools.clone(),
-            spawn_only: self.spawn_only.clone(),
-            spawn_only_messages: self.spawn_only_messages.clone(),
-        }
-    }
-
-    // ── Deferred tool activation ──────────────────────────────────────
-
-    /// Mark tools as deferred (hidden from specs until activated).
-    /// Call during setup before wrapping in Arc.
-    pub fn defer(&mut self, names: impl IntoIterator<Item = String>) {
-        let deferred = self.deferred.get_mut().unwrap_or_else(|e| e.into_inner());
-        for name in names {
-            if self.tools.contains_key(&name) {
-                deferred.insert(name);
-            }
-        }
-        self.invalidate_cache();
-    }
-
-    /// Defer all tools in a named group (e.g. "group:web").
-    pub fn defer_group(&mut self, group: &str) {
-        if let Some(info) = policy::tool_group_info(group) {
-            let deferred = self.deferred.get_mut().unwrap_or_else(|e| e.into_inner());
-            for &tool in info.tools {
-                if self.tools.contains_key(tool) {
-                    deferred.insert(tool.to_string());
-                }
-            }
-            self.invalidate_cache();
-        }
-    }
-
-    /// Activate a deferred tool group or individual tool. Works through `&self`
-    /// (interior mutability) so it can be called during the agent loop via Arc.
-    /// Returns the names of tools that were activated.
-    pub fn activate(&self, group_or_name: &str) -> Vec<String> {
-        let mut deferred = self.deferred.lock().unwrap_or_else(|e| e.into_inner());
-        let mut activated = Vec::new();
-
-        if let Some(info) = policy::tool_group_info(group_or_name) {
-            for &tool in info.tools {
-                // Never activate spawn_only tools — they must be used via spawn subagent
-                if !self.spawn_only.contains(tool) && deferred.remove(tool) {
-                    activated.push(tool.to_string());
-                }
-            }
-        } else if !self.spawn_only.contains(group_or_name) && deferred.remove(group_or_name) {
-            activated.push(group_or_name.to_string());
-        }
-
-        if !activated.is_empty() {
-            self.invalidate_cache_shared();
-        }
-        activated
-    }
-
-    /// Returns info about currently deferred tool groups for the activate_tools tool.
-    pub fn deferred_groups(&self) -> Vec<(String, String, usize)> {
-        let deferred = self.deferred.lock().unwrap_or_else(|e| e.into_inner());
-        if deferred.is_empty() {
-            return Vec::new();
-        }
-
-        let mut groups = Vec::new();
-        for info in policy::TOOL_GROUPS {
-            let count = info.tools.iter().filter(|&&t| deferred.contains(t)).count();
-            if count > 0 {
-                groups.push((info.name.to_string(), info.description.to_string(), count));
-            }
-        }
-
-        // Also list individually deferred tools not in any group
-        let grouped: HashSet<&str> = policy::TOOL_GROUPS
-            .iter()
-            .flat_map(|g| g.tools.iter().copied())
-            .collect();
-        for name in deferred.iter() {
-            if !grouped.contains(name.as_str()) {
-                groups.push((name.clone(), "Plugin tool".to_string(), 1));
-            }
-        }
-        groups
-    }
-
-    /// Whether any tools are currently deferred.
-    pub fn has_deferred(&self) -> bool {
-        !self
-            .deferred
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .is_empty()
-    }
-
-    // ── LRU auto-eviction ────────────────────────────────────────────
-
-    /// Mark a set of tool names as "base" — never auto-evicted.
-    pub fn set_base_tools(&mut self, names: impl IntoIterator<Item = impl Into<String>>) {
-        self.lifecycle
-            .get_mut()
-            .unwrap_or_else(|e| e.into_inner())
-            .set_base_tools(names);
-    }
-
-    /// Add more tool names to the base set (extends, does not replace).
-    pub fn add_base_tools(&mut self, names: impl IntoIterator<Item = impl Into<String>>) {
-        self.lifecycle
-            .get_mut()
-            .unwrap_or_else(|e| e.into_inner())
-            .add_base_tools(names);
-    }
-
-    /// Record that a tool was used (called from execute()).
-    fn record_usage(&self, name: &str) {
-        self.lifecycle
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .record_usage(name);
-    }
-
-    /// Advance the iteration counter. Called before each LLM call.
-    pub fn tick(&self) {
-        self.lifecycle
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .tick();
-    }
-
-    /// Auto-evict idle non-base tools if active count exceeds threshold.
-    /// Returns the names of evicted tools (for logging).
-    ///
-    /// Lock ordering: lifecycle → deferred (consistent with record_usage
-    /// which only takes lifecycle, never both).
-    pub fn auto_evict(&self) -> Vec<String> {
-        // 1. Compute eviction candidates (lifecycle lock only)
-        let to_evict = {
-            let lifecycle = self.lifecycle.lock().unwrap_or_else(|e| e.into_inner());
-            let deferred = self.deferred.lock().unwrap_or_else(|e| e.into_inner());
-            let active: Vec<&str> = self
-                .tools
-                .keys()
-                .filter(|n| !deferred.contains(n.as_str()))
-                .map(|n| n.as_str())
-                .collect();
-            lifecycle.find_evictable(&active)
-            // Both locks dropped here
-        };
-
-        if to_evict.is_empty() {
-            return Vec::new();
-        }
-
-        // 2. Apply evictions (deferred lock only)
-        {
-            let mut deferred = self.deferred.lock().unwrap_or_else(|e| e.into_inner());
-            for name in &to_evict {
-                deferred.insert(name.clone());
-            }
-        }
-        self.invalidate_cache_shared();
-
-        to_evict
-    }
-
-    // ── Cache management ────────────────────────────────────────────
-
-    /// Clear the cached specs (called by mutation methods with &mut self).
-    fn invalidate_cache(&mut self) {
-        // &mut self guarantees exclusive access, so get_mut() bypasses the mutex.
-        *self
-            .cached_specs
-            .get_mut()
-            .unwrap_or_else(|e| e.into_inner()) = None;
-    }
-
-    /// Clear the cached specs through &self (for interior-mutability callers).
-    fn invalidate_cache_shared(&self) {
-        *self.cached_specs.lock().unwrap_or_else(|e| e.into_inner()) = None;
-    }
-
-    /// Execute a tool by name.
-    ///
-    /// Respects provider policy: tools hidden from `specs()` are also blocked
-    /// from execution. This prevents an LLM from calling tools it shouldn't
-    /// have access to.
-    pub async fn execute(&self, name: &str, args: &serde_json::Value) -> Result<ToolResult> {
-        if let Some(ref policy) = self.provider_policy {
-            if !policy.is_allowed(name) {
-                eyre::bail!("tool '{}' denied by provider policy", name);
-            }
-        }
-
-        // Auto-activate deferred tools on first use — no need for the LLM
-        // to call activate_tools first. This prevents the retry loop where
-        // the LLM keeps calling a deferred tool and getting errors.
-        {
-            let deferred = self.deferred.lock().unwrap_or_else(|e| e.into_inner());
-            if deferred.contains(name) {
-                drop(deferred);
-                // Find which group this tool belongs to and activate the whole group
-                let group = policy::TOOL_GROUPS
-                    .iter()
-                    .find(|g| g.tools.contains(&name))
-                    .map(|g| g.name);
-                if let Some(group_name) = group {
-                    let activated = self.activate(group_name);
-                    tracing::info!(
-                        tool = name,
-                        group = group_name,
-                        activated = %activated.join(", "),
-                        "auto-activated deferred tool on first use"
-                    );
-                } else {
-                    // Not in any group — activate individually
-                    let mut deferred = self.deferred.lock().unwrap_or_else(|e| e.into_inner());
-                    deferred.remove(name);
-                    drop(deferred);
-                    self.invalidate_cache_shared();
-                    tracing::info!(tool = name, "auto-activated deferred tool (no group)");
-                }
-            }
-        }
-
-        // Reject oversized arguments (1 MB limit).
-        // Use a counting writer to avoid allocating the full serialized string,
-        // which could OOM on deeply nested JSON before the check triggers.
-        const MAX_ARGS_SIZE: usize = 1_048_576;
-        let args_size = estimate_json_size(args);
-        if args_size > MAX_ARGS_SIZE {
-            eyre::bail!(
-                "tool '{}' arguments too large: ~{} bytes (max {})",
-                name,
-                args_size,
-                MAX_ARGS_SIZE
-            );
-        }
-
-        let tool = self
-            .tools
-            .get(name)
-            .ok_or_else(|| eyre::eyre!("unknown tool: {}", name))?;
-
-        // Track usage for LRU auto-eviction
-        self.record_usage(name);
-
-        tool.execute(args).await
-    }
-}
+// Tool registry (extracted to its own module)
+mod registry;
+pub use registry::ToolRegistry;
 
 // Tool policy
 pub mod policy;
@@ -698,41 +237,6 @@ pub use git::GitTool;
 pub use code_structure::CodeStructureTool;
 
 use std::path::{Component, Path};
-
-use crate::sandbox::{NoSandbox, Sandbox};
-
-/// Estimate the serialized JSON size without allocating.
-/// Walks the serde_json::Value tree recursively, counting bytes.
-fn estimate_json_size(value: &serde_json::Value) -> usize {
-    match value {
-        serde_json::Value::Null => 4,
-        serde_json::Value::Bool(b) => {
-            if *b {
-                4
-            } else {
-                5
-            }
-        }
-        serde_json::Value::Number(n) => n.to_string().len(),
-        serde_json::Value::String(s) => {
-            let escapes = s
-                .bytes()
-                .filter(|&b| matches!(b, b'"' | b'\\' | b'\n' | b'\r' | b'\t'))
-                .count();
-            s.len() + escapes + 2 // content + escape overheads + quotes
-        }
-        serde_json::Value::Array(arr) => {
-            2 + arr.iter().map(estimate_json_size).sum::<usize>() + arr.len().saturating_sub(1) // commas
-        }
-        serde_json::Value::Object(obj) => {
-            2 + obj
-                .iter()
-                .map(|(k, v)| k.len() + 3 + estimate_json_size(v))
-                .sum::<usize>()
-                + obj.len().saturating_sub(1) // commas
-        }
-    }
-}
 
 /// Resolve a user-provided path, ensuring it stays within base_dir.
 ///
@@ -894,181 +398,6 @@ pub fn file_io_error(e: std::io::Error, display_path: &str) -> ToolResult {
     }
 }
 
-impl ToolRegistry {
-    /// Create a registry with built-in tools for the given working directory.
-    pub fn with_builtins(cwd: impl AsRef<Path>) -> Self {
-        Self::with_builtins_and_sandbox(cwd, Box::new(NoSandbox))
-    }
-
-    /// Create a registry with built-in tools and a custom sandbox for shell commands.
-    pub fn with_builtins_and_sandbox(cwd: impl AsRef<Path>, sandbox: Box<dyn Sandbox>) -> Self {
-        let cwd = cwd.as_ref();
-        let mut registry = Self::new();
-        registry.register(ShellTool::new(cwd).with_sandbox(sandbox));
-        registry.register(ReadFileTool::new(cwd));
-        registry.register(DiffEditTool::new(cwd));
-        registry.register(EditFileTool::new(cwd));
-        registry.register(WriteFileTool::new(cwd));
-        registry.register(GlobTool::new(cwd));
-        registry.register(GrepTool::new(cwd));
-        registry.register(ListDirTool::new(cwd));
-        registry.register(WebSearchTool::new());
-        registry.register(WebFetchTool::new());
-        registry.register(BrowserTool::new());
-        #[cfg(feature = "git")]
-        registry.register(GitTool::new(cwd));
-        #[cfg(feature = "ast")]
-        registry.register(CodeStructureTool::new(cwd));
-        registry
-    }
-
-    /// Tool names that are bound to a working directory (cwd / base_dir).
-    /// Used by `rebind_cwd()` to re-register these tools with a new workspace path.
-    pub const CWD_BOUND_TOOLS: &'static [&'static str] = &[
-        "shell",
-        "read_file",
-        "write_file",
-        "edit_file",
-        "diff_edit",
-        "glob",
-        "grep",
-        "list_dir",
-        #[cfg(feature = "git")]
-        "git",
-        #[cfg(feature = "ast")]
-        "code_structure",
-    ];
-
-    /// Create a copy of this registry with all cwd-bound tools re-registered
-    /// to use a new working directory and sandbox. Non-cwd tools (web_search,
-    /// web_fetch, browser, MCP, plugins, etc.) are preserved via Arc cloning.
-    pub fn rebind_cwd(&self, cwd: impl AsRef<Path>, sandbox: Box<dyn Sandbox>) -> Self {
-        let cwd = cwd.as_ref();
-        // Clone everything except cwd-bound tools
-        let mut registry = self.snapshot_excluding(Self::CWD_BOUND_TOOLS);
-        // Re-register cwd-bound tools with the new workspace
-        registry.register(ShellTool::new(cwd).with_sandbox(sandbox));
-        registry.register(ReadFileTool::new(cwd));
-        registry.register(DiffEditTool::new(cwd));
-        registry.register(EditFileTool::new(cwd));
-        registry.register(WriteFileTool::new(cwd));
-        registry.register(GlobTool::new(cwd));
-        registry.register(GrepTool::new(cwd));
-        registry.register(ListDirTool::new(cwd));
-        #[cfg(feature = "git")]
-        registry.register(GitTool::new(cwd));
-        #[cfg(feature = "ast")]
-        registry.register(CodeStructureTool::new(cwd));
-        registry
-    }
-
-    /// Re-bind all plugin tools to a new work directory.
-    ///
-    /// Creates copies of each `PluginTool` with the given work_dir so that
-    /// per-session output (e.g. voice profiles) lands inside the user's
-    /// workspace where the agent's sandboxed tools can access it.
-    pub fn rebind_plugin_work_dirs(&mut self, work_dir: &Path) {
-        use crate::plugins::PluginTool;
-        let replacements: Vec<_> = self
-            .tools
-            .iter()
-            .filter_map(|(name, tool)| {
-                tool.as_any()
-                    .downcast_ref::<PluginTool>()
-                    .map(|pt| (name.clone(), pt.clone_with_work_dir(work_dir.to_path_buf())))
-            })
-            .collect();
-        for (name, new_tool) in replacements {
-            self.tools.insert(name, Arc::new(new_tool));
-        }
-    }
-
-    /// Re-register builtin configurable tools with a ToolConfigStore.
-    ///
-    /// Tools already registered by `with_builtins_and_sandbox()` are replaced
-    /// with config-aware instances. Also registers the `configure_tool` tool.
-    pub fn inject_tool_config(&mut self, config: Arc<ToolConfigStore>) {
-        if self.tools.contains_key("web_search") {
-            self.register(WebSearchTool::new().with_config(config.clone()));
-        }
-        if self.tools.contains_key("web_fetch") {
-            self.register(WebFetchTool::new().with_config(config.clone()));
-        }
-        if self.tools.contains_key("browser") {
-            self.register(BrowserTool::new().with_config(config.clone()));
-        }
-        self.register(ConfigureToolTool::new(config));
-    }
-}
-
-#[cfg(test)]
-mod estimate_tests {
-    use super::*;
-
-    #[test]
-    fn test_null() {
-        assert_eq!(estimate_json_size(&serde_json::Value::Null), 4);
-    }
-
-    #[test]
-    fn test_bool() {
-        assert_eq!(estimate_json_size(&serde_json::json!(true)), 4);
-        assert_eq!(estimate_json_size(&serde_json::json!(false)), 5);
-    }
-
-    #[test]
-    fn test_number() {
-        assert_eq!(estimate_json_size(&serde_json::json!(42)), 2);
-        assert_eq!(estimate_json_size(&serde_json::json!(2.72)), 4);
-    }
-
-    #[test]
-    fn test_string_simple() {
-        // "hello" -> 5 chars + 2 quotes = 7
-        assert_eq!(estimate_json_size(&serde_json::json!("hello")), 7);
-    }
-
-    #[test]
-    fn test_string_with_escapes() {
-        // "a\"b" has 3 chars + 1 escape overhead + 2 quotes = 6
-        assert_eq!(estimate_json_size(&serde_json::json!("a\"b")), 6);
-        // "a\nb" has 3 chars + 1 escape + 2 quotes = 6
-        assert_eq!(estimate_json_size(&serde_json::json!("a\nb")), 6);
-    }
-
-    #[test]
-    fn test_empty_array() {
-        assert_eq!(estimate_json_size(&serde_json::json!([])), 2);
-    }
-
-    #[test]
-    fn test_array_with_elements() {
-        // [1,2,3] = 2 brackets + 3 numbers (1+1+1) + 2 commas = 7
-        assert_eq!(estimate_json_size(&serde_json::json!([1, 2, 3])), 7);
-    }
-
-    #[test]
-    fn test_empty_object() {
-        assert_eq!(estimate_json_size(&serde_json::json!({})), 2);
-    }
-
-    #[test]
-    fn test_object_with_fields() {
-        // {"a":1} = 2 braces + key(1) + 3 (quotes+colon) + value(1) = 7
-        let v = serde_json::json!({"a": 1});
-        assert_eq!(estimate_json_size(&v), 7);
-    }
-
-    #[test]
-    fn test_nested_structure() {
-        let v = serde_json::json!({"x": [1, 2]});
-        // Outer: 2 + key(1+3) + inner array
-        // Inner array: 2 + 1 + 1 + 1 comma = 5
-        // Total: 2 + 4 + 5 = 11
-        assert_eq!(estimate_json_size(&v), 11);
-    }
-}
-
 #[cfg(test)]
 mod nofollow_tests {
     use super::*;
@@ -1211,17 +540,12 @@ mod path_tests {
     /// resolve_path must block access to other profiles' directories.
     #[test]
     fn test_resolve_blocks_cross_profile_access() {
-        // Simulate per-profile isolation: cwd = ~/.octos/profiles/alice/data
         let base = Path::new("/home/user/.octos/profiles/alice/data");
 
-        // Trying to read another profile's data via traversal
         assert!(resolve_path(base, "../../bob/data/sessions/secret").is_err());
         assert!(resolve_path(base, "../../../profiles/bob/data/episodes.db").is_err());
-
-        // Trying to reach shared octos home
         assert!(resolve_path(base, "../../../skills/evil-skill/main").is_err());
 
-        // Valid access within own data dir
         assert!(resolve_path(base, "skills/my-skill/main").is_ok());
         assert!(resolve_path(base, "sessions/chat-123.json").is_ok());
         assert!(resolve_path(base, "skill-output/report.pdf").is_ok());
@@ -1230,8 +554,6 @@ mod path_tests {
     #[test]
     fn test_resolve_rejects_empty_path() {
         let base = Path::new("/home/user/project");
-        // Empty string should resolve to the base dir itself, which is valid
-        // (it's not a traversal). Verify it doesn't panic.
         let result = resolve_path(base, "");
         assert!(result.is_ok());
         assert_eq!(result.unwrap(), PathBuf::from("/home/user/project"));
@@ -1240,13 +562,7 @@ mod path_tests {
     #[test]
     fn test_resolve_rejects_null_byte() {
         let base = Path::new("/home/user/project");
-        // Null bytes in paths are invalid on Unix — verify resolve_path
-        // handles them without panicking. The OS will reject the path.
         let result = resolve_path(base, "file\0.txt");
-        // On Unix, Path::new("file\0.txt") is valid at the Rust level but
-        // normalize_path will produce a path that starts_with base, so it
-        // passes. The actual OS call (open/stat) would reject it.
-        // The key security property: it doesn't escape the base directory.
         if let Ok(p) = &result {
             assert!(p.starts_with(base));
         }
@@ -1255,453 +571,9 @@ mod path_tests {
     #[test]
     fn test_resolve_rejects_windows_separators() {
         let base = Path::new("/home/user/project");
-        // Backslash is a valid filename char on Unix, not a separator.
-        // Verify it doesn't enable traversal.
         let result = resolve_path(base, "..\\..\\etc\\passwd");
-        // On Unix this is a single filename component "..\..\etc\passwd",
-        // which stays inside base.
         if let Ok(p) = &result {
             assert!(p.starts_with(base));
         }
-    }
-}
-
-/// Integration test: verifies that `rebind_cwd` produces a registry where
-/// file tools reject paths outside the new working directory.
-#[cfg(test)]
-mod cwd_isolation_tests {
-    use super::*;
-    use crate::sandbox::NoSandbox;
-
-    #[tokio::test]
-    async fn test_rebind_cwd_file_tools_reject_outside_paths() {
-        // Create initial registry with a broad cwd
-        let broad_cwd = std::path::Path::new("/tmp");
-        let registry = ToolRegistry::with_builtins_and_sandbox(broad_cwd, Box::new(NoSandbox));
-
-        // Now rebind to a narrow cwd (simulating per-profile isolation)
-        let narrow_cwd = tempfile::tempdir().expect("create temp dir");
-        let narrow = narrow_cwd.path();
-        let rebound = registry.rebind_cwd(narrow, Box::new(NoSandbox));
-
-        // Create a file inside the narrow cwd so we can test reads
-        let inside_file = narrow.join("allowed.txt");
-        std::fs::write(&inside_file, "hello").expect("write test file");
-
-        // read_file inside cwd should succeed
-        let result = rebound
-            .execute("read_file", &serde_json::json!({"path": "allowed.txt"}))
-            .await;
-        assert!(result.is_ok(), "read inside narrow cwd should work");
-        let tr = result.unwrap();
-        assert!(tr.success, "read_file should succeed: {}", tr.output);
-
-        // read_file with traversal outside cwd should fail
-        let result = rebound
-            .execute(
-                "read_file",
-                &serde_json::json!({"path": "../../etc/passwd"}),
-            )
-            .await;
-        assert!(result.is_ok(), "should not return transport error");
-        let tr = result.unwrap();
-        assert!(
-            !tr.success,
-            "read_file with traversal should be rejected: {}",
-            tr.output
-        );
-
-        // write_file outside cwd should fail
-        let result = rebound
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "../escape.txt",
-                    "content": "pwned"
-                }),
-            )
-            .await;
-        assert!(result.is_ok());
-        let tr = result.unwrap();
-        assert!(
-            !tr.success,
-            "write_file outside narrow cwd should be rejected: {}",
-            tr.output
-        );
-
-        // glob inside cwd should work
-        let result = rebound
-            .execute("glob", &serde_json::json!({"pattern": "*.txt"}))
-            .await;
-        assert!(result.is_ok());
-        let tr = result.unwrap();
-        assert!(tr.success, "glob inside cwd should work: {}", tr.output);
-
-        // list_dir inside cwd should work
-        let result = rebound
-            .execute("list_dir", &serde_json::json!({"path": "."}))
-            .await;
-        assert!(result.is_ok());
-        let tr = result.unwrap();
-        assert!(tr.success, "list_dir inside cwd should work: {}", tr.output);
-
-        // list_dir with traversal should fail
-        let result = rebound
-            .execute("list_dir", &serde_json::json!({"path": "../../"}))
-            .await;
-        assert!(result.is_ok());
-        let tr = result.unwrap();
-        assert!(
-            !tr.success,
-            "list_dir with traversal should be rejected: {}",
-            tr.output
-        );
-    }
-
-    #[tokio::test]
-    async fn test_rebind_cwd_preserves_non_cwd_tools() {
-        let initial_cwd = tempfile::tempdir().expect("create temp dir");
-        let registry =
-            ToolRegistry::with_builtins_and_sandbox(initial_cwd.path(), Box::new(NoSandbox));
-
-        let new_cwd = tempfile::tempdir().expect("create temp dir");
-        let rebound = registry.rebind_cwd(new_cwd.path(), Box::new(NoSandbox));
-
-        // Non-cwd tools should still be present
-        assert!(
-            rebound.get("web_fetch").is_some(),
-            "web_fetch should survive rebind"
-        );
-        assert!(
-            rebound.get("web_search").is_some(),
-            "web_search should survive rebind"
-        );
-
-        // CWD-bound tools should also be present (re-registered)
-        assert!(
-            rebound.get("read_file").is_some(),
-            "read_file should be re-registered"
-        );
-        assert!(
-            rebound.get("shell").is_some(),
-            "shell should be re-registered"
-        );
-        assert!(
-            rebound.get("write_file").is_some(),
-            "write_file should be re-registered"
-        );
-    }
-}
-
-#[cfg(test)]
-mod lifecycle_tests {
-    use super::*;
-    use std::path::PathBuf;
-    use std::sync::Arc;
-
-    fn make_registry(max_active: usize, idle_threshold: u32) -> ToolRegistry {
-        let mut reg = ToolRegistry::with_builtins(PathBuf::from("/tmp"));
-        {
-            let lc = reg.lifecycle.get_mut().unwrap();
-            lc.max_active = max_active;
-            lc.idle_threshold = idle_threshold;
-        }
-        reg
-    }
-
-    fn active_tool_names(reg: &ToolRegistry) -> Vec<String> {
-        let mut names: Vec<String> = reg.specs().iter().map(|s| s.name.clone()).collect();
-        names.sort();
-        names
-    }
-
-    fn deferred_tool_names(reg: &ToolRegistry) -> Vec<String> {
-        let deferred = reg.deferred.lock().unwrap();
-        let mut names: Vec<String> = deferred.iter().cloned().collect();
-        names.sort();
-        names
-    }
-
-    // ── Scenario 1: Basic eviction ──────────────────────────────────
-
-    #[test]
-    fn idle_tools_evicted_when_over_threshold() {
-        // Set up: max 3 active, idle after 2 iterations
-        let mut reg = make_registry(3, 2);
-
-        // Mark read_file and write_file as base (never evict)
-        reg.set_base_tools(["read_file", "write_file"]);
-
-        let initial_count = reg.specs().len();
-        println!("Initial active tools: {initial_count}");
-        assert!(initial_count > 3, "builtins should exceed threshold");
-
-        // Simulate 3 iterations of using only read_file and write_file
-        for _ in 0..3 {
-            reg.tick();
-            reg.record_usage("read_file");
-            reg.record_usage("write_file");
-        }
-
-        // Now evict — tools not used in 2+ iterations should be evicted
-        let evicted = reg.auto_evict();
-        println!("Evicted: {evicted:?}");
-        assert!(!evicted.is_empty(), "should evict idle tools");
-
-        // Base tools should still be active
-        let active = active_tool_names(&reg);
-        assert!(
-            active.contains(&"read_file".to_string()),
-            "base tool read_file must survive"
-        );
-        assert!(
-            active.contains(&"write_file".to_string()),
-            "base tool write_file must survive"
-        );
-
-        // Evicted tools should be in deferred
-        let deferred = deferred_tool_names(&reg);
-        for name in &evicted {
-            assert!(
-                deferred.contains(name),
-                "{name} should be deferred after eviction"
-            );
-        }
-
-        println!(
-            "After eviction — active: {}, deferred: {}",
-            active.len(),
-            deferred.len()
-        );
-        assert!(active.len() <= 3, "should be at or under threshold");
-    }
-
-    // ── Scenario 2: Recently used tools survive eviction ────────────
-
-    #[test]
-    fn recently_used_tools_not_evicted() {
-        let mut reg = make_registry(3, 2);
-        reg.set_base_tools(["read_file"]);
-
-        // Use shell heavily, leave others idle
-        for _ in 0..3 {
-            reg.tick();
-            reg.record_usage("read_file");
-            reg.record_usage("shell");
-        }
-
-        let evicted = reg.auto_evict();
-        println!("Evicted: {evicted:?}");
-
-        // shell was used every iteration — should NOT be evicted
-        assert!(
-            !evicted.contains(&"shell".to_string()),
-            "recently used 'shell' should not be evicted"
-        );
-
-        let active = active_tool_names(&reg);
-        assert!(
-            active.contains(&"shell".to_string()),
-            "shell must remain active"
-        );
-    }
-
-    // ── Scenario 3: Deferred tool activated then used ───────────────
-
-    #[tokio::test]
-    async fn activated_tool_gets_usage_tracking() {
-        let mut reg = make_registry(3, 2);
-        reg.set_base_tools(["read_file", "write_file"]);
-
-        // Simulate iterations to trigger eviction
-        for _ in 0..3 {
-            reg.tick();
-            reg.record_usage("read_file");
-            reg.record_usage("write_file");
-        }
-        let evicted = reg.auto_evict();
-        println!("First eviction: {evicted:?}");
-        assert!(!evicted.is_empty());
-
-        // Verify shell is deferred
-        let deferred = deferred_tool_names(&reg);
-        let shell_was_evicted = deferred.contains(&"shell".to_string());
-        println!("shell deferred: {shell_was_evicted}");
-
-        if shell_was_evicted {
-            // Re-activate shell (simulates activate_tools call)
-            let activated = reg.activate("group:runtime");
-            println!("Activated: {activated:?}");
-            assert!(activated.contains(&"shell".to_string()));
-
-            // Now use shell
-            reg.tick();
-            reg.record_usage("shell");
-
-            // Shell should not be evicted next time — it's freshly used
-            let evicted2 = reg.auto_evict();
-            assert!(
-                !evicted2.contains(&"shell".to_string()),
-                "freshly used shell should survive eviction"
-            );
-            println!("Second eviction (shell survived): {evicted2:?}");
-        }
-    }
-
-    // ── Scenario 4: Base tools never evicted even when idle ─────────
-
-    #[test]
-    fn base_tools_never_evicted() {
-        let mut reg = make_registry(2, 1);
-        reg.set_base_tools(["read_file", "write_file", "shell"]);
-
-        // Never use any tool, just tick
-        for _ in 0..5 {
-            reg.tick();
-        }
-
-        let evicted = reg.auto_evict();
-        println!("Evicted: {evicted:?}");
-
-        // Base tools must survive regardless of idleness
-        for name in &["read_file", "write_file", "shell"] {
-            assert!(
-                !evicted.contains(&name.to_string()),
-                "base tool {name} must never be evicted"
-            );
-        }
-    }
-
-    // ── Scenario 5: Stalest tool evicted first ──────────────────────
-
-    #[test]
-    fn stalest_evicted_first() {
-        let mut reg = make_registry(5, 2);
-        reg.set_base_tools(["read_file"]);
-
-        // Iteration 1: use several tools
-        reg.tick();
-        reg.record_usage("read_file");
-        reg.record_usage("shell");
-        reg.record_usage("write_file");
-        reg.record_usage("edit_file");
-        reg.record_usage("glob");
-        reg.record_usage("grep");
-        reg.record_usage("list_dir");
-
-        // Iteration 2-4: only use shell and write_file
-        for _ in 0..3 {
-            reg.tick();
-            reg.record_usage("shell");
-            reg.record_usage("write_file");
-        }
-
-        let evicted = reg.auto_evict();
-        println!("Evicted: {evicted:?}");
-
-        // edit_file, glob, grep, list_dir were last used at iteration 1
-        // shell and write_file were used at iteration 4
-        // Stalest (edit_file, glob, grep, list_dir) should be evicted first
-        if !evicted.is_empty() {
-            assert!(
-                !evicted.contains(&"shell".to_string()),
-                "shell (iter 4) should survive over stale tools"
-            );
-            assert!(
-                !evicted.contains(&"write_file".to_string()),
-                "write_file (iter 4) should survive over stale tools"
-            );
-        }
-    }
-
-    // ── Scenario 6: No eviction when under threshold ────────────────
-
-    #[test]
-    fn no_eviction_when_under_threshold() {
-        let mut reg = make_registry(100, 1); // Very high threshold
-
-        for _ in 0..5 {
-            reg.tick();
-        }
-
-        let evicted = reg.auto_evict();
-        assert!(evicted.is_empty(), "should not evict when under threshold");
-    }
-
-    // ── Scenario 7: Full session lifecycle ──────────────────────────
-
-    #[tokio::test]
-    async fn full_session_lifecycle() {
-        let mut reg = make_registry(5, 3);
-        reg.set_base_tools(["read_file", "write_file"]);
-
-        println!("=== Turn 1: Research query ===");
-        reg.tick();
-        reg.record_usage("read_file");
-        reg.record_usage("shell");
-        let active = active_tool_names(&reg);
-        println!("Active ({}): {:?}", active.len(), active);
-
-        println!("\n=== Turns 2-4: Only using read/write ===");
-        for i in 2..=4 {
-            reg.tick();
-            reg.record_usage("read_file");
-            reg.record_usage("write_file");
-            let evicted = reg.auto_evict();
-            if !evicted.is_empty() {
-                println!("Turn {i} evicted: {evicted:?}");
-            }
-        }
-        let active = active_tool_names(&reg);
-        let deferred = deferred_tool_names(&reg);
-        println!(
-            "After turn 4 — active: {}, deferred: {}",
-            active.len(),
-            deferred.len()
-        );
-
-        println!("\n=== Turn 5: Need shell again — re-activate ===");
-        if deferred.contains(&"shell".to_string()) {
-            let activated = reg.activate("group:runtime");
-            println!("Activated: {activated:?}");
-        }
-        reg.tick();
-        reg.record_usage("shell");
-        let active = active_tool_names(&reg);
-        println!(
-            "Active after re-activation ({}): {:?}",
-            active.len(),
-            active
-        );
-        assert!(
-            active.contains(&"shell".to_string()),
-            "shell should be active again"
-        );
-
-        println!("\n=== Turn 6-8: Use shell, others go idle ===");
-        for i in 6..=8 {
-            reg.tick();
-            reg.record_usage("read_file");
-            reg.record_usage("shell");
-            let evicted = reg.auto_evict();
-            if !evicted.is_empty() {
-                println!("Turn {i} evicted: {evicted:?}");
-            }
-        }
-        let active = active_tool_names(&reg);
-        let deferred = deferred_tool_names(&reg);
-        println!(
-            "\nFinal state — active: {}, deferred: {}",
-            active.len(),
-            deferred.len()
-        );
-        println!("Active: {:?}", active);
-        println!("Deferred: {:?}", deferred);
-
-        // Base tools must always be active
-        assert!(active.contains(&"read_file".to_string()));
-        assert!(active.contains(&"write_file".to_string()));
-        // Shell was recently used — should survive
-        assert!(active.contains(&"shell".to_string()));
     }
 }
